Moved providers into own file
